Terminator more elegant
